+++ src/LanguageServer/Console.fs
@@ -9,6 +9,11 @@ module Console =
     /// Concat two lines into one line for command help, so they won't be separated by other empty line
     let inline (<+>) line1 line2 = sprintf "%s\n%s" line1 line2
 
+    [<RequireQualifiedAccess>]
+    type WatchSubdirs =
+        | Yes
+        | No
+
     type FileOrDir =
         | File of string
         | Dir of string * string list
@@ -39,43 +44,83 @@ module Console =
             | File file -> [ file ]
             | Dir (_, files) -> files
 
-    open Tuc.Domain
-    open ErrorHandling
+        let watch = function
+            | File file -> file, WatchSubdirs.No
+            | Dir (dir, _) -> dir, WatchSubdirs.Yes
 
-    let parseDomain (input, output) domain =
-        domain
-        |> FileOrDir.files
-        |> List.map (Parser.parse output)
-
-    let checkDomain (input, output) domain =
-        result {
-            let parsedDomains =
-                domain
-                |> parseDomain (input, output)
-
-            let! resolvedTypes =
-                parsedDomains
-                |> Resolver.resolve output
-                |> Result.mapError UnresolvedTypes
-
-            let! domainTypes =
-                resolvedTypes
-                |> Checker.check output
-                |> Result.mapError UndefinedTypes
-
-            return domainTypes
+    [<RequireQualifiedAccess>]
+    module internal Watch =
+        let private runForever = async {
+            while true do
+                do! Async.Sleep 1000
         }
 
-    let showParseDomainError output = function
-        | UnresolvedTypes unresolvedTypes ->
-            unresolvedTypes
-            |> List.map (TypeName.value >> List.singleton)
-            |> output.Options (sprintf "Unresolved types [%d]:" (unresolvedTypes |> List.length))
+        let watch output watchSubdirs execute (path, filter) = async {
+            let includeSubDirs =
+                match watchSubdirs with
+                | WatchSubdirs.Yes -> true
+                | WatchSubdirs.No -> false
 
-            output.Error "You have to solve unresolved types first.\n"
-        | UndefinedTypes undefinedTypes ->
-            undefinedTypes
-            |> List.map (TypeName.value >> List.singleton)
-            |> output.Options (sprintf "Undefined types [%d]:" (undefinedTypes |> List.length))
+            let pathDir, fileName =
+                if path |> Directory.Exists then path, None
+                elif path |> File.Exists then Path.GetDirectoryName(path), Some path
+                else failwithf "Path %A is invalid." path
 
-            output.Error "You have to define all types first.\n"
+            use watcher =
+                new FileSystemWatcher(
+                    Path = pathDir,
+                    EnableRaisingEvents = true,
+                    IncludeSubdirectories = includeSubDirs
+                )
+
+            watcher.Filters.Add(filter)
+
+            match fileName with
+            | Some fileName ->
+                watcher.Filters.Add(fileName)
+            | _ -> ()
+
+            if output.IsDebug() then
+                sprintf "<c:gray>[Watch]</c> Path: <c:cyan>%s</c> | Filters: <c:yellow>%s</c> | With subdirs: <c:magenta>%A</c>"
+                    path
+                    (watcher.Filters |> String.concat "; ")
+                    includeSubDirs
+                |> output.Message
+
+            watcher.NotifyFilter <- watcher.NotifyFilter ||| NotifyFilters.LastWrite
+            watcher.SynchronizingObject <- null
+
+            let notifyWatch () =
+                path
+                |> sprintf "<c:gray>[Watch]</c> Watching path <c:dark-yellow>%A</c> (Press <c:yellow>ctrl + c</c> to stop) ...\n"
+                |> output.Message
+
+            let executeOnWatch event =
+                if output.IsDebug() then output.Message <| sprintf "<c:gray>[Watch]</c> Source %s." event
+
+                output.Message "<c:gray>[Watch]</c> Executing ...\n"
+
+                try execute()
+                with e -> output.Error <| sprintf "%A" e
+
+                notifyWatch ()
+
+            watcher.Changed.Add(fun _ -> executeOnWatch "changed")
+            watcher.Created.Add(fun _ -> executeOnWatch "created")
+            watcher.Deleted.Add(fun _ -> executeOnWatch "deleted")
+            watcher.Renamed.Add(fun _ -> executeOnWatch "renamed")
+
+            if output.IsVerbose() then
+                output.Message <| sprintf "<c:gray>[Watch]</c> Enabled for %A" path
+
+            notifyWatch()
+
+            do! runForever
+        }
+
+        let executeAndWaitForWatch output execute = async {
+            try execute()
+            with e -> output.Error <| sprintf "%A" e
+
+            do! runForever
+        }