+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Parse only Domain `.fsx` files (`*Domain.fsx`)
+- Watch and resolve Domain `.fsx` files internally
+- Move `DomainTypes` to `State`
+- Add `tuc/domainResolved` notification
 
 ## 1.0.0 - 2020-11-13
 - Initial implementation